feat: schedules get their own Command-Center tab + a boss-room calendar
The recurring-missions list was buried at the bottom of the monitor
tab's long scroll. This pulls it into a first-class SCHEDULES tab and
adds a physical way in from the office floor.

- SchedulesTab: self-contained tab (mission list, on/off toggle, delete, add form, live "last fired / next" via onMissionsUpdated). The monitor tab loses ~150 lines of scroll.
- A wall calendar hangs in the CEO office (tear-off month page: red banner, day grid, today circled). Clicking it selects Michael and opens the schedules tab.
- New store one-shot requestCommandCenterTab / ccTabRequest (seq-keyed) drives the tab switch from the floor prop; the panel consumes it.
- New "clock" pixel icon for the tab.
